refactor: move metadata to own folder
