+++ src/data-workspace/data-entry-cell/data-entry-cell.js
@@ -4,8 +4,11 @@ import PropTypes from 'prop-types'
 import React, { useState } from 'react'
 import { useField } from 'react-final-form'
 import { useContextSelection } from '../../context-selection/index.js'
-import { useMetadata } from '../metadata-context.js'
-import { getCategoryComboById, getDataSetById } from '../selectors.js'
+import { useMetadata } from '../../metadata/index.js'
+import {
+    getCategoryComboById,
+    getDataSetById,
+} from '../../metadata/selectors.js'
 import styles from './data-entry-cell.module.css'
 import { getValidatorByValueType } from './field-validation.js'
 import { useDataValueMutation } from './use-data-value-mutation.js'
