+++ src/data-workspace/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
+import { MetadataContextProvider } from '../metadata/index.js'
 import { DataWorkspace as SimpleDataWorkspace } from './data-workspace.js'
-import { MetadataContextProvider } from './metadata-context.js'
 
 export const DataWorkspace = () => (
     <MetadataContextProvider>
