+++ src/shared/index.js
@@ -1,3 +1,4 @@
 export * from './error-message/index.js'
 export * from './fixed-periods/index.js'
 export * from './period/index.js'
+export * from './utils.js'
